### 2023-10-01 #5909 @editorialbot
[PRE REVIEW]: Opyrability: A Python package for process operability analysis
**Submitting author:** @victoraalves (<a href="http://orcid.org/0009-0006-5308-9717">Victor Alves</a>)
**Repository:** https://github.com/CODES-group/opyrability
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.3
