### 2023-10-01 #5909 @kyleniemeyer
Hello @kramea, would you be able to review this submission to JOSS?
