+++ paper/references.bib
@@ -22,7 +22,7 @@ pages = "728 - 733",
 year = "2015",
 note = "9th IFAC Symposium on Advanced Control of Chemical Processes ADCHEM 2015",
 issn = "2405-8963",
-doi = "https://doi.org/10.1016/j.ifacol.2015.09.055",
+doi = "10.1016/j.ifacol.2015.09.055",
 url = "http://www.sciencedirect.com/science/article/pii/S2405896315011362",
 author = "Juan C. Carrasco and Fernando V. Lima",
 keywords = "operability, nonlinear systems, membrane reactor, direct methane aromatization, system analysis, constraints",
@@ -97,7 +97,7 @@ volume = "105",
 pages = "246 - 258",
 year = "2017",
 issn = "0098-1354",
-doi = "https://doi.org/10.1016/j.compchemeng.2016.12.010",
+doi = "10.1016/j.compchemeng.2016.12.010",
 url = "http://www.sciencedirect.com/science/article/pii/S009813541630415X",
 author = "Juan C. Carrasco and Fernando V. Lima",
 keywords = "Process intensification, Bilevel optimization, Operability, Modular systems, Natural gas utilization",
@@ -128,7 +128,7 @@ pages = "179 - 184",
 year = "2015",
 booktitle = "12th International Symposium on Process Systems Engineering and 25th European Symposium on Computer Aided Process Engineering",
 issn = "1570-7946",
-doi = "https://doi.org/10.1016/B978-0-444-63578-5.50025-6",
+doi = "10.1016/B978-0-444-63578-5.50025-6",
 url = "http://www.sciencedirect.com/science/article/pii/B9780444635785500256",
 author = "Natalia Quirante and Juan Javaloyes and Rubén Ruiz-Femenia and José A. Caballero",
 keywords = "process optimization, design, kriging algorithm, modular simulators",
@@ -446,7 +446,8 @@ year = {2010}
   number={5},
   pages={433--440},
   year={1943},
-  publisher={American Society of Mechanical Engineers}
+  publisher={American Society of Mechanical Engineers},
+  doi={10.1115/1.4018788}
 }
 
 
