### 2023-10-03 #5909 @victoraalves
@kyleniemeyer Based on the JOSS list, some other reviewers can be, in case the others don't respond: Gilberto M. Xavier, Andy Shahidehpour, Mark E Fuller. I focused my search on people related to Chemical Engineering & Python development. Needless to say I declare no COIs. I don't know any of these people.
