+++ tests/cstr_new.py
@@ -1,0 +1,181 @@
+import jax.numpy as np
+from scipy.optimize import root
+import matplotlib.pyplot as plt
+from opyrability import AIS2AOS_map
+from jax import jacrev
+
+
+beta = 0.35
+xi = 1.00
+gamma = 20.00
+phi = 0.11
+sigma_s = 0.44
+sigma_b = 0.06
+q0 = 1.00
+qc = 1.00
+x10 = 1.00
+x20 = 0.00
+x30 = -0.05
+
+
+# beta = 1.33
+# xi = 0.6
+# gamma = 25.18
+# phi = 19.26
+# sigma_s = 91.42
+# sigma_b = 11.43
+# q0 = 1.00
+# x10 = 1.00
+# x20 =  -0.12
+# x30 = - 0.12
+# qc  = 6.56
+
+
+
+# Define cstr function
+def cstr(x, AIS):
+    
+
+    fx1 = np.exp((gamma*x)/(1+x))
+    
+    term1 =  q0*(x20 - x)
+    term2 =  (beta*phi*fx1*AIS[1]*q0*x10) / (q0 + phi*fx1*AIS[1])
+    term3 = ((sigma_s*AIS[1] + sigma_b)*AIS[0]*qc)/(qc*AIS[0] + xi*(sigma_s*AIS[1] + sigma_b))
+    term4 = (x - x30)
+    
+    g_x1 = term1 + term2 - (term3*term4)
+    
+    return g_x1
+
+
+# jac_cstr_raw=jacrev(cstr)
+
+# def jac_cstr(x, *args):
+#     return np.array(jac_cstr_raw(x, *args))
+
+# Define m function
+# def m(u):
+    
+    
+    
+#     solution = root(cstr, x0 = 0.1, args=u, method='hybr', tol=1e-16)
+ 
+#     if solution.success is True:
+#         Temp_dimensionless = solution.x
+#         fx1 = np.exp((gamma*Temp_dimensionless)/(1 + Temp_dimensionless))
+#         xA_dimensionless =  ((q0*x10)/ (q0 + phi*fx1*u[1]))
+#         # conversion = ((x10 - xA_dimensionless)/ x10)
+#         conversion = 1 - ( xA_dimensionless / x10)
+#     else:
+#         Temp_dimensionless = np.nan
+#         fx1 = np.nan
+#         xA_dimensionless = np.nan
+#         conversion = np.nan
+        
+    
+    
+    
+#     return np.array([Temp_dimensionless, conversion]).reshape(2,)
+
+
+def m(u):
+    try:
+        solution = root(cstr, x0=0.1, args=u, method='hybr', tol=1e-16)
+        
+        if not solution.success:
+            # If the first attempt failed, try with the last iteration as the initial guess
+            solution = root(cstr, x0=solution.x, args=u, method='hybr', tol=1e-16)
+        
+        # If the second attempt is also unsuccessful, raise an exception to move to the except block
+        if not solution.success:
+            raise ValueError("Solver failed to converge.")
+        
+        Temp_dimensionless = solution.x
+        fx1 = np.exp((gamma*Temp_dimensionless)/(1 + Temp_dimensionless))
+        xA_dimensionless = ((q0*x10) / (q0 + phi*fx1*u[1]))
+        conversion = 1 - (xA_dimensionless / x10)
+        
+    except ValueError:
+        Temp_dimensionless = np.nan
+        fx1 = np.nan
+        xA_dimensionless = np.nan
+        conversion = np.nan
+
+    return np.array([Temp_dimensionless, conversion]).reshape(2,)
+
+# # Initialize the u values
+# u_values = np.array([
[…]
