+++ tests/cstr_new.py
@@ -4,7 +4,11 @@ import matplotlib.pyplot as plt
 from opyrability import AIS2AOS_map
 from jax import jacrev
 
+from jax.config import config
+config.update("jax_enable_x64", True)
 
+
+# Values from Ray #
 beta = 0.35
 xi = 1.00
 gamma = 20.00
@@ -18,17 +22,41 @@ x20 = 0.00
 x30 = -0.05
 
 
-# beta = 1.33
-# xi = 0.6
-# gamma = 25.18
-# phi = 19.26
-# sigma_s = 91.42
-# sigma_b = 11.43
-# q0 = 1.00
-# x10 = 1.00
-# x20 =  -0.12
-# x30 = - 0.12
-# qc  = 6.56
+
+# ------------------------ Values from Luyben ------------------------------ #
+
+def findA(E,k):
+    R,T = 1.99, 599.67
+    
+    
+    return k / np.exp(-E/(R*T))
+
+E1, E2 = 30000, 15000
+k1, k2 = 0.5, 0.05
+
+Aa =  findA(E1,k1)
+Ab =  findA(E2, k2)
+
+
+beta = 1.33
+xi = 0.6
+gamma = 25.18
+phi = 19.26
+sigma_s = 91.42
+sigma_b = 11.43
+q0 = 1.00
+x10 = 1.00
+x20 =  -0.12
+x30 = - 0.12
+qc  = 6.56
+
+dH1 = 20000
+dH2 = 20000
+eta1 = E1/E2
+eta2 = dH1/dH2
+
+phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
+
 
 
 
@@ -48,10 +76,21 @@ def cstr(x, AIS):
     return g_x1
 
 
-# jac_cstr_raw=jacrev(cstr)
 
-# def jac_cstr(x, *args):
-#     return np.array(jac_cstr_raw(x, *args))
+def cstr2(x, AIS):
+    
+    fx3 = np.exp((gamma*x)/(1+x))
+    
+    term1 = q0*(x30 - x)
+    term2 = phi*fx3*AIS[1]*beta
+    term3 = (q0*x10)/(q0 + phi*fx3*AIS[1])
+    term4 = 
+
+
+jac_cstr1_raw=jacrev(cstr)
+
+def jac_cstr(x, *args):
+    return np.array(jac_cstr1_raw(x, *args))
 
 # Define m function
 # def m(u):
@@ -76,15 +115,18 @@ def cstr(x, AIS):
     
     
 #     return np.array([Temp_dimensionless, conversion]).reshape(2,)
-
+ftol=1e-8
 
 def m(u):
     try:
-        solution = root(cstr, x0=0.1, args=u, method='hybr', tol=1e-16)
+        options = {'xtol': 1e-8, 'ftol': ftol }
+        solution = root(cstr, x0=0.1, args=u, method='lm', jac=jac_cstr, 
+                        tol=ftol, options = options)
         
         if not solution.success:
             # If the first attempt failed, try with the last iteration as the initial guess
-            solution = root(cstr, x0=solution.x, args=u, method='hybr', tol=1e-16)
+            solution = root(cstr, x0=solution.x, args=u, method='lm', jac=jac_cstr, 
+                            tol=ftol, options = options)
         
         # If the second attempt is also unsuccessful, raise an exception to move to the except block
         if not solution.success:
@@ -94,16 +136,45 @@ def m(u):
         fx1 = np.exp((gamma*Temp_dimensionless)/(1 + Temp_dimensionless))
         xA_dimensionless = ((q0*x10) / (q0 + phi*fx1*u[1]))
[…]
