### 2023-10-11 da0be9b
continuation of cstr2 model
