+++ tests/cstr_new.py
@@ -1,25 +1,25 @@
 import jax.numpy as np
-from scipy.optimize import root
+from scipy.optimize import root, root_scalar
 import matplotlib.pyplot as plt
 from opyrability import AIS2AOS_map
-from jax import jacrev
+from jax import jacrev, grad, jit, jacfwd, hessian
 
 from jax.config import config
 config.update("jax_enable_x64", True)
 
 
 # Values from Ray #
-beta = 0.35
-xi = 1.00
-gamma = 20.00
-phi = 0.11
-sigma_s = 0.44
-sigma_b = 0.06
-q0 = 1.00
-qc = 1.00
-x10 = 1.00
-x20 = 0.00
-x30 = -0.05
+# beta = 0.35
+# xi = 1.00
+# gamma = 20.00
+# phi = 0.11
+# sigma_s = 0.44
+# sigma_b = 0.06
+# q0 = 1.00
+# qc = 1.00
+# x10 = 1.00
+# x20 = 0.00
+# x30 = -0.05
 
 
 
@@ -54,12 +54,16 @@ dH1 = 20000
 dH2 = 20000
 eta1 = E1/E2
 eta2 = dH1/dH2
+Tr  = 599.67
+Tc0 = 529.67
+
+x40 =  (Tc0 - Tr)/ Tr
 
 phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
 
 
 
-
+@jit
 # Define cstr function
 def cstr(x, AIS):
     
@@ -84,14 +88,40 @@ def cstr2(x, AIS):
     term1 = q0*(x30 - x)
     term2 = phi*fx3*AIS[1]*beta
     term3 = (q0*x10)/(q0 + phi*fx3*AIS[1])
-    term4 = 
+    term4 = (eta2*phi2*(fx3**(eta1-1)))/(q0 + phi*phi2*(fx3**(eta1))*AIS[1])
+    term5 = (q0*x20 + (phi*fx3*AIS[1]*q0*x10)/(q0 + phi*fx3*AIS[1]))
+    term6 = (sigma_s*AIS[1] + sigma_b)*qc*AIS[0]*(x - x40)
+    term7 = (qc*AIS[0] + (sigma_s*AIS[1] +  sigma_b)*xi)
+    
+    g = term1 + term2*(term3 + (term4*term5)) + (term6/term7)
+    
+    return g
+
+
 
 
 jac_cstr1_raw=jacrev(cstr)
+hess_cstr_raw = jacfwd(jacrev(cstr2))
 
+grad_cstr2_raw =  jacrev(cstr2)
+
+hess_cstr2_raw = jacfwd(jacrev(cstr2))
+
+
+
+@jit
 def jac_cstr(x, *args):
     return np.array(jac_cstr1_raw(x, *args))
 
+@jit
+def hess_cstr(x, *args):
+    return np.array(hess_cstr_raw(x, *args))
+@jit
+def grad_cstr2(x, *args):
+    return np.array(grad_cstr2_raw(x, *args))
+@jit
+def hess_cstr2(x, *args):
+    return np.array(hess_cstr2_raw(x, *args))
 # Define m function
 # def m(u):
     
@@ -147,6 +177,42 @@ def m(u):
         jacket_temp = np.nan
 
     return np.array([Temp_dimensionless, conversion]).reshape(2,)
+
+
+def m_cstr2(u):
+    solution = root_scalar(cstr2, x0=0.1, args=u, method='halley',
+                           fprime=grad_cstr2, fprime2 = hess_cstr2,
+                           bracket= (0, 1))
+    
+    
+    Temp_dimensionless = solution.root
[…]
