### 2023-10-11 #5909 @kyleniemeyer
Thanks @gmxavier! We have [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) written up you can look through. Briefly, once we find a second reviewer, I will start a separate review issue, where you will generate a checklist of items to work through.

You'll need to download and install the software, and inspect the repository
and submitted paper for key elements. Any feedback you have for the authors can be given as a comment in the review issue, or as an issue at the package's repository.
