+++ tests/cstr_new.py
@@ -1,5 +1,5 @@
 import jax.numpy as np
-from scipy.optimize import root, root_scalar
+from scipy.optimize import root, root_scalar, fsolve
 import matplotlib.pyplot as plt
 from opyrability import AIS2AOS_map
 from jax import jacrev, grad, jit, jacfwd, hessian
@@ -9,57 +9,57 @@ config.update("jax_enable_x64", True)
 
 
 # Values from Ray #
-# beta = 0.35
-# xi = 1.00
-# gamma = 20.00
-# phi = 0.11
-# sigma_s = 0.44
-# sigma_b = 0.06
-# q0 = 1.00
-# qc = 1.00
-# x10 = 1.00
-# x20 = 0.00
-# x30 = -0.05
+beta = 0.35
+xi = 1.00
+gamma = 20.00
+phi = 0.11
+sigma_s = 0.44
+sigma_b = 0.06
+q0 = 1.00
+qc = 1.00
+x10 = 1.00
+x20 = 0.00
+x30 = -0.05
 
 
 
 # ------------------------ Values from Luyben ------------------------------ #
 
-def findA(E,k):
-    R,T = 1.99, 599.67
+# def findA(E,k):
+#     R,T = 1.99, 599.67
     
     
-    return k / np.exp(-E/(R*T))
+#     return k / np.exp(-E/(R*T))
 
-E1, E2 = 30000, 15000
-k1, k2 = 0.5, 0.05
+# E1, E2 = 30000, 15000
+# k1, k2 = 0.5, 0.05
 
-Aa =  findA(E1,k1)
-Ab =  findA(E2, k2)
+# Aa =  findA(E1,k1)
+# Ab =  findA(E2, k2)
 
 
-beta = 1.33
-xi = 0.6
-gamma = 25.18
-phi = 19.26
-sigma_s = 91.42
-sigma_b = 11.43
-q0 = 1.00
-x10 = 1.00
-x20 =  -0.12
-x30 = - 0.12
-qc  = 6.56
+# beta = 1.33
+# xi = 0.6
+# gamma = 25.18
+# phi = 19.26
+# sigma_s = 91.42
+# sigma_b = 11.43
+# q0 = 1.00
+# x10 = 1.00
+# x20 =  -0.12
+# x30 = - 0.12
+# qc  = 6.56
 
-dH1 = 20000
-dH2 = 20000
-eta1 = E1/E2
-eta2 = dH1/dH2
-Tr  = 599.67
-Tc0 = 529.67
+# dH1 = 20000
+# dH2 = 20000
+# eta1 = E1/E2
+# eta2 = dH1/dH2
+# Tr  = 599.67
+# Tc0 = 529.67
 
-x40 =  (Tc0 - Tr)/ Tr
+# x40 =  (Tc0 - Tr)/ Tr
 
-phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
+# phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
 
 
 
@@ -68,11 +68,11 @@ phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
 def cstr(x, AIS):
     
 
-    fx1 = np.exp((gamma*x)/(1+x))
+    fx2 = np.exp((gamma*x)/(1+x))
     
     term1 =  q0*(x20 - x)
-    term2 =  (beta*phi*fx1*AIS[1]*q0*x10) / (q0 + phi*fx1*AIS[1])
[…]
