### 2023-10-13 5e7ecc7
secant method works better for cstr. (10-12-23)
