### 2023-10-13 1f5445b
multiplicity region mapped using opencv.
