+++ tests/cstr_new.py
@@ -3,6 +3,7 @@ from scipy.optimize import root, root_scalar, fsolve
 import matplotlib.pyplot as plt
 from opyrability import AIS2AOS_map
 from jax import jacrev, grad, jit, jacfwd, hessian
+import cyipopt
 
 from jax.config import config
 config.update("jax_enable_x64", True)
@@ -67,16 +68,16 @@ x30 = -0.05
 # Define cstr function
 def cstr(x, AIS):
     
-
     fx2 = np.exp((gamma*x)/(1+x))
-    
-    term1 =  q0*(x20 - x)
-    term2 =  (beta*phi*fx2*AIS[1]*q0*x10) / (q0 + phi*fx2*AIS[1])
-    term3 = (((sigma_s*AIS[1] + sigma_b)*AIS[0]*qc)/(qc*AIS[0] + xi*(sigma_s*AIS[1] + sigma_b)))
+
+    term1 = q0*(x20 - x)
+    term2 = (beta*phi*fx2*AIS[1]*q0*x10) / (q0 + phi*fx2*AIS[1])
+    term3 = (((sigma_s*AIS[1] + sigma_b)*AIS[0]*qc) /
+             (qc*AIS[0] + xi*(sigma_s*AIS[1] + sigma_b)))
     term4 = (x - x30)
-    
+
     g_x1 = term1 + term2 - (term3*term4)
-    
+
     return g_x1
 
 
@@ -84,17 +85,17 @@ def cstr(x, AIS):
 def cstr2(x, AIS):
     
     fx3 = np.exp((gamma*x)/(1+x))
-    
+
     term1 = q0*(x30 - x)
     term2 = phi*fx3*AIS[1]*beta
     term3 = (q0*x10)/(q0 + phi*fx3*AIS[1])
     term4 = (eta2*phi2*(fx3**(eta1-1)))/(q0 + phi*phi2*(fx3**(eta1))*AIS[1])
     term5 = (q0*x20 + (phi*fx3*AIS[1]*q0*x10)/(q0 + phi*fx3*AIS[1]))
     term6 = (sigma_s*AIS[1] + sigma_b)*qc*AIS[0]*(x - x40)
-    term7 = (qc*AIS[0] + (sigma_s*AIS[1] +  sigma_b)*xi)
-    
+    term7 = (qc*AIS[0] + (sigma_s*AIS[1] + sigma_b)*xi)
+
     g = term1 + term2*(term3 + (term4*term5)) + (term6/term7)
-    
+
     return g
 
 
@@ -104,7 +105,6 @@ jac_cstr1_raw=grad(cstr)
 hess_cstr_raw = grad(grad(cstr))
 
 grad_cstr2_raw =  grad(cstr2)
-
 hess_cstr2_raw = grad(grad(cstr2))
 
 
@@ -122,71 +122,35 @@ def grad_cstr2(x, *args):
 @jit
 def hess_cstr2(x, *args):
     return np.array(hess_cstr2_raw(x, *args))
-# Define m function
-# def m(u):
-    
-    
-    
-#     solution = root(cstr, x0 = 0.1, args=u, method='hybr', tol=1e-16)
- 
-#     if solution.success is True:
-#         Temp_dimensionless = solution.x
-#         fx1 = np.exp((gamma*Temp_dimensionless)/(1 + Temp_dimensionless))
-#         xA_dimensionless =  ((q0*x10)/ (q0 + phi*fx1*u[1]))
-#         # conversion = ((x10 - xA_dimensionless)/ x10)
-#         conversion = 1 - ( xA_dimensionless / x10)
-#     else:
-#         Temp_dimensionless = np.nan
-#         fx1 = np.nan
-#         xA_dimensionless = np.nan
-#         conversion = np.nan
-        
-    
-    
-    
[…]
