### 2023-10-16 6019ea7
cstr implicit map start
