+++ tests/cstr_new.py
@@ -1,135 +1,162 @@
 import jax.numpy as np
 from scipy.optimize import root, root_scalar, fsolve
 import matplotlib.pyplot as plt
-from opyrability import AIS2AOS_map
+from opyrability import AIS2AOS_map, create_grid
 from jax import jacrev, grad, jit, jacfwd, hessian
 import cyipopt
+import pandas as pd
 
 from jax.config import config
 config.update("jax_enable_x64", True)
 
 
-# Values from Ray #
-beta = 0.35
-xi = 1.00
-gamma = 20.00
-phi = 0.11
-sigma_s = 0.44
-sigma_b = 0.06
-q0 = 1.00
-qc = 1.00
-x10 = 1.00
-x20 = 0.00
-x30 = -0.05
+# ---------------------------- Set I of parameters -------------------------- #
+# Values from Ray (1982) applicable for CSTR Model 1 
+#                            from Subramanian (2003)
+
+beta_1    = 0.35
+xi_1      = 1.00
+gamma_1   = 20.00
+phi_1     = 0.11
+sigma_s_1 = 0.44
+sigma_b_1 = 0.06
+q0_1      = 1.00
+qc_1      = 1.00
+x10_1     = 1.00
+x20_1     = 0.00
+x30_1     = -0.05
 
 
+# ---------------------------- Set II of parameters ------------------------- #
+# Values from Luyben (1993) applicable for CSTR Model 1 
+#                            from Subramanian (2003)
 
-# ------------------------ Values from Luyben ------------------------------ #
-
-# def findA(E,k):
-#     R,T = 1.99, 599.67
+def findA(E,k):
+    R,T = 1.99, 599.67
     
     
-#     return k / np.exp(-E/(R*T))
+    return k / np.exp(-E/(R*T))
 
-# E1, E2 = 30000, 15000
-# k1, k2 = 0.5, 0.05
+E1, E2 = 30000, 15000
+k1, k2 = 0.5, 0.05
 
-# Aa =  findA(E1,k1)
-# Ab =  findA(E2, k2)
+Aa      =  findA(E1,k1)
+Ab      =  findA(E2, k2)
+beta    = 1.33
+xi      = 0.6
+gamma   = 25.18
+phi     = 19.26
+sigma_s = 91.42
+sigma_b = 11.43
+q0      = 1.00
+x10     = 1.00
+x20     = -0.12
+x30     = - 0.12
+qc      = 6.56
+# I had to go to Luyben '93 to find these (original source).
+dH1     = 20000
+dH2     = 20000
+eta1    = E1/E2
+eta2    = dH1/dH2
+Tr      = 599.67
+Tc0     = 529.67
 
-
-# beta = 1.33
-# xi = 0.6
-# gamma = 25.18
-# phi = 19.26
-# sigma_s = 91.42
-# sigma_b = 11.43
-# q0 = 1.00
-# x10 = 1.00
-# x20 =  -0.12
-# x30 = - 0.12
-# qc  = 6.56
-
-# dH1 = 20000
-# dH2 = 20000
-# eta1 = E1/E2
-# eta2 = dH1/dH2
-# Tr  = 599.67
-# Tc0 = 529.67
-
-# x40 =  (Tc0 - Tr)/ Tr
-
-# phi2 = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
+# x40, phi2 values were not given but it can be calculated.
+x40     =  (Tc0 - Tr)/ Tr
+phi2    = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
 
 
-
-@jit
-# Define cstr function
+# ---------------------------- 1d CSTR functions ---------------------------- #
[…]
