### 2023-10-19 4ed276a
cstr almost done 18-10-23
