+++ tests/cstr_implicit.py
@@ -1,0 +1,538 @@
+import jax
+import jax.numpy as np
+from scipy.optimize import root, root_scalar, fsolve
+import matplotlib.pyplot as plt
+from opyrability import AIS2AOS_map, create_grid
+from jax import jacrev, grad, jit, jacfwd, hessian, vmap
+import cyipopt
+import pandas as pd
+from jaxopt import ScipyRootFinding, Broyden, Bisection
+from opyrability import implicit_map as imap
+
+from jax.config import config
+config.update("jax_enable_x64", True)
+
+def generate_edge_points(bounds, n_points):
+    """
+    Generates boundary points on the edges of a hypercube defined by the given bounds.
+    
+    Parameters:
+    - bounds: numpy array with shape (n_dimensions, 2) where each row is 
+    [lower_bound, upper_bound] for a dimension.
+    - n_points: total number of points to generate.
+    
+    Returns:
+    - edge_points: numpy array with the boundary points.
+    """
+
+    n_dims = bounds.shape[0]
+    points_per_dim = int(round(n_points ** (1/n_dims)))
+
+    edge_sets = []
+    for i in range(n_dims):
+        for bound in bounds[i]:
+            slice_points = [np.linspace(bounds[j, 0], bounds[j, 1], points_per_dim) if j != i else np.array([bound])
+                            for j in range(n_dims)]
+            mesh = np.array(np.meshgrid(*slice_points)).T.reshape(-1, n_dims)
+            edge_sets.append(mesh)
+
+    edge_points = np.vstack(edge_sets)
+    return edge_points
+
+
+class NonlinearSystemSolver:
+    def __init__(self, funcs, args):
+        self.funcs = funcs
+        self.args = args
+        # Compute the Jacobian of the vector function using jacrev
+        self.jac_func = jacrev(self.vector_func)
+        self.hess_func = hessian(self.objective)
+    
+    def vector_func(self, x):
+        """Vector function representing the system of equations."""
+        return np.array([f(x, self.args) for f in self.funcs])
+    
+    def objective(self, x):
+        """Objective function: sum of squares of residuals."""
+        residuals = np.array([f(x, self.args) for f in self.funcs])
+        return np.sum(residuals**2)
+    
+    def gradient(self, x):
+        """Calculate gradient using the Jacobian."""
+        # Evaluate the Jacobian at the current x
+        J = self.jac_func(x)
+        # Compute the gradient as J^T * f(x)
+        return J.T @ self.vector_func(x)
+    
+    def hessian(self, x, lagrange, obj_factor):
+        """Calculate Hessian of the objective function."""
+        # For systems of equations without constraints, the Hessian of the Lagrangian 
+        # is just obj_factor times the Hessian of the objective.
+        H = self.hess_func(x)
+        return obj_factor * H
+    
+    def solve(self, x0):
+        """Solve the system using IPOPT."""
+        n = len(x0)
+        x_L = np.ones(n) * -1.0e0
+        x_U = np.ones(n) * 1.0e0
+        problem = cyipopt.Problem(n=n, m=0, problem_obj=self, lb=x_L, ub=x_U)
[…]
