### 2023-10-20 95d7531
cstr - jupyter notebook
