+++ tests/cstr_implicit.py
@@ -3,11 +3,12 @@ import jax.numpy as np
 from scipy.optimize import root, root_scalar, fsolve
 import matplotlib.pyplot as plt
 from opyrability import AIS2AOS_map, create_grid
-from jax import jacrev, grad, jit, jacfwd, hessian, vmap
+from jax import jacrev, jit, vmap
 import cyipopt
 import pandas as pd
-from jaxopt import ScipyRootFinding, Broyden, Bisection
+from jaxopt import ScipyRootFinding, Broyden
 from opyrability import implicit_map as imap
+import time
 
 from jax.config import config
 config.update("jax_enable_x64", True)
@@ -38,48 +39,6 @@ def generate_edge_points(bounds, n_points):
 
     edge_points = np.vstack(edge_sets)
     return edge_points
-
-
-class NonlinearSystemSolver:
-    def __init__(self, funcs, args):
-        self.funcs = funcs
-        self.args = args
-        # Compute the Jacobian of the vector function using jacrev
-        self.jac_func = jacrev(self.vector_func)
-        self.hess_func = hessian(self.objective)
-    
-    def vector_func(self, x):
-        """Vector function representing the system of equations."""
-        return np.array([f(x, self.args) for f in self.funcs])
-    
-    def objective(self, x):
-        """Objective function: sum of squares of residuals."""
-        residuals = np.array([f(x, self.args) for f in self.funcs])
-        return np.sum(residuals**2)
-    
-    def gradient(self, x):
-        """Calculate gradient using the Jacobian."""
-        # Evaluate the Jacobian at the current x
-        J = self.jac_func(x)
-        # Compute the gradient as J^T * f(x)
-        return J.T @ self.vector_func(x)
-    
-    def hessian(self, x, lagrange, obj_factor):
-        """Calculate Hessian of the objective function."""
-        # For systems of equations without constraints, the Hessian of the Lagrangian 
-        # is just obj_factor times the Hessian of the objective.
-        H = self.hess_func(x)
-        return obj_factor * H
-    
-    def solve(self, x0):
-        """Solve the system using IPOPT."""
-        n = len(x0)
-        x_L = np.ones(n) * -1.0e0
-        x_U = np.ones(n) * 1.0e0
-        problem = cyipopt.Problem(n=n, m=0, problem_obj=self, lb=x_L, ub=x_U)
-        problem.addOption('print_level', 0)
-        x, info = problem.solve(x0)
-        return x
 
 
 # ---------------------------- Set I of parameters -------------------------- #
@@ -140,7 +99,7 @@ phi2    = (Ab * np.exp(-gamma*(eta1 - 1 ))) / Aa
 
 # ---------------------------- 1d CSTR functions ---------------------------- #
 # Define cstr functions
-@jit
+# @jit
 def cstr(x, AIS):
     
     beta_1    = 0.35
@@ -172,7 +131,7 @@ def cstr(x, AIS):
     Eqn = Eqn.astype(float)
     return Eqn
 
-
+# @jit
 def cstr2(x, AIS):
     
     fx3 = np.exp((gamma*x)/(1+x))
@@ -203,10 +162,15 @@ def cstr2(x, AIS):
 def cstr_equation(Temp_dimensionless, AIS):
     return cstr(Temp_dimensionless, AIS)
 
-rtol=1e-12
+rtol=1e-6
 xtol= 1e-6
 gtol = 1e-6
 
+
+# Using JaxOPT (Differentiable Root-finding from Google)
+cstr_solver = Broyden(cstr_equation, tol=rtol, maxls=1000, jit=True)
+
[…]
