+++ tests/load_test.py
@@ -1,0 +1,3 @@
+import numpy as npp
+
+npp.load('montecarlo.npz')
