+++ tests/old_implicit.py
@@ -20,6 +20,433 @@ import matplotlib.pyplot as plt
 
 from opyrability import implicit_map as imap
 
+
+
+# %% Test area! REMOVE BEFORE RELEASE!!!!
+def shower_implicit(u,y):
+    d = jnp.zeros(2)
+    y0_aux = (u[0]+u[1])
+    
+    y0_aux= jnp.where(y0_aux <= 1e-9, 1e-9, y0_aux)
+    
+    
+    LHS1 = y[0] - (u[0]+u[1])
+    LHS2 = y[1] - (u[0]*(60+d[0])+u[1]*(120+d[1]))/(y0_aux)
+    
+    
+    # LHS2_AX = y[1] - (60+120)/2
+    LHS2_AX = y[1] - (60+120)/2
+    LHS2 = jnp.where(y0_aux <= 1e-9, LHS2_AX, LHS2)
+    
+    # if y[0]!=0:
+    #     LHS2 = y[1] - (u[0]*(60+d[0])+u[1]*(120+d[1]))/(u[0]+u[1])
+    # else:
+        
+    
+    return jnp.array([LHS1, LHS2])
+
+
+
+def shower(u):
+    y = np.zeros(2)
+    d = jnp.zeros(2)
+    y[0] = (u[0]+u[1])
+    # LHS2 = y[1] - (u[0]*(60+d[0])+u[1]*(120+d[1]))/(u[0]+u[1])
+    if y[0]!=0:
+        y[1] = (u[0]*(60+d[0])+u[1]*(120+d[1]))/(u[0]+u[1])
+    else:
+        y[1] = (60+120)/2
+    
+    return jnp.array(y)
+
+def FF1(u):
+    y = np.zeros(2)
+    # y[0] = u[1]**2*u[0]
+    y[0] = u[0] - 2*u[1]
+    y[1] = 3*u[0] + 4*u[1]
+    return jnp.array(y)
+
+def FF1_implicit(u,y):
+    # LHS0 = y[0] - u[1]**2*u[0]
+    LHS0 = y[0] - (u[0] - 2*u[1])
+    LHS1 = y[1] - (3*u[0] + 4*u[1])
+    return jnp.array([LHS0, LHS1])
+
+
+
+from matplotlib.patches import Polygon
+
+def makeplot(AOS_poly):
+    fig, ax = plt.subplots()
+    for i in range(len(AOS_poly)):
+        x = AOS_poly[i][0]
+        y = AOS_poly[i][1]
+        order = np.argsort(np.arctan2(y - y.mean(), x - x.mean()))
+        A = AOS_poly[i].T
+        polygon = Polygon(A[order] ,facecolor='tab:blue', edgecolor='black', linewidth=1)
+        ax.add_patch(polygon)
+    ax.autoscale_view()
+
+#%% Test DMA-MR inverse
+# DOS_bound = np.array([[22.4, 22.8],
+#                     [39.4, 40.0]])
+
+# DOSresolution = [10, 10]
+
+# output_init = np.array([20.0, 0.9])
+
+# DOS, DIS, DOS_poly, DIS_poly = imap(F_DMA_MR_eqn,
+#                                     DOS_bound,
+#                                     DOSresolution,
+#                                     output_init,
+#                                     direction='inverse')
+
+# %% Test shower forward
+# AIS_bound = np.array([[0.1, 10.0],
+#                     [0.1, 10.0]])
+
+# # AIS_bound =  np.array([[5.0, 10.0],
+# #                       [80.0, 100.0]])
+
+# AISresolution = [5, 5]
+
+# output_init = np.array([0.0, 10.0])
+
+
[…]

[…]
