+++ tests/load_test.py
@@ -1,3 +1,0 @@
-import numpy as npp
-
-npp.load('montecarlo.npz')
