### 2023-10-31 #5966 @gmxavier
@victoraalves About the example usage. The first example (shower) ran flawless, but the same didn’t occur for the second one (DMA-MR). It failed, at least for me, at this command:

```
fDIS, fDOS, convergence = nlp_based_approach(dma_mr_design,
                                             DOS_bounds,
                                             DOS_resolution,
                                             u0,
                                             lb,
                                             ub,
                                             constr=(con),
                                             method='ipopt',
                                             plot=True,
                                             ad=True,
                                             warmstart=True)

You have selected automatic differentiation as a method for obtaining higher-order data (Jacobians/Hessian),
 Make sure your process model is JAX-compatible implementation-wise.
  0%|          | 0/100 [00:00<?, ?it/s]
[<ipython-input-15-257d67bb9950>](https://localhost:8080/#) in <cell line: 1>()
----> 1 fDIS, fDOS, convergence = nlp_based_approach(dma_mr_design,
      2                                              DOS_bounds,
      3                                              DOS_resolution,
      4                                              u0,
      5                                              lb,

6 frames
[/usr/local/lib/python3.10/site-packages/jax/_src/api.py](https://localhost:8080/#) in _check_input_dtype_revderiv(name, holomorphic, allow_int, x)
    769       dtypes.issubdtype(aval.dtype, np.bool_)):
    770     if not allow_int:
--> 771       raise TypeError(f"{name} requires real- or complex-valued inputs (input dtype "
    772                       f"that is a sub-dtype of np.inexact), but got {aval.dtype.name}. "
    773                       "If you want to use Boolean- or integer-valued inputs, use vjp "

TypeError: jacrev requires real- or complex-valued inputs (input dtype that is a sub-dtype of np.inexact), but got int32. If you want to use Boolean- or integer-valued inputs, use vjp or set allow_int to True.
```

Could you help with this?

Note: I'm running my checks using Colab. I've installed the package using the following commands:

```
!pip install -q condacolab
import condacolab
condacolab.install()
!conda install -c codes-group -c conda-forge opyrability
```
