+++ pyproject.toml
@@ -30,7 +30,7 @@ exclude = ["*.tests", "*.tests.*", "tests.*", "tests", "*.obsolete", "*.obsolete
 
 [tool.poetry.dependencies]
 python = "^3.9.0"
-numpy = "^1.24.1"
+numpy = ">=1.24.1,<1.26.0"
 scipy = "^1.10.0"
 tqdm = "^4.64.1"
 matplotlib = "^3.6.2"
