### 2023-11-01 #5966 @victoraalves
@gmxavier Thank you for the feedback! The problem was related to jax on linux (colab) not being able to recognize some datatypes within numpy arrays. This is easily fixed by making sure that the jax.numpy arrays contain floats:

Instead of:

```
DOS_bounds = np.array([[15,25],
                       [35,45]])

# Discretization Grid - 10x10 grid for DOS.
DOS_resolution =  [10, 10]

# Lower and upper bounds of AIS design variables.
lb = np.array([10,  0.1])
ub = np.array([300, 2])

# Initial estimate for NLP.
u0 = np.array([50, 1])
```

It should be:

```
DOS_bounds = np.array([[15.00,25.00],
                       [35.00,45.00]])

# Discretization Grid - 10x10 grid for DOS.
DOS_resolution =  [10, 10]

# Lower and upper bounds of AIS design variables.
lb = np.array([10.00,  0.1])
ub = np.array([300.00, 2.00])

# Initial estimate for NLP.
u0 = np.array([50.00, 1.00])
```

Note the dot (.) after the numbers for the arrays of _lb, ub and u0._

**_But this is kind of boring forcing the user to ensure floats with "." or ".00"._**  There is a more elgant, **definitive** solution which is checking if the datatypes are integers and then setting them to float32/float64. I have implemented this (see below) and it works now:

https://github.com/CODES-group/opyrability/blob/94840595b38353af4396f67d6a527ff72b566529/src/opyrability.py#L752-L755

This will make the code work for now (irrespective of being in a colab environment or not) but I added a quick datatype check to ensure that the numpy arrays are type=float (see above). This never happened under Windows and on Linux via WSL. This might be related to jax running in the Colab's virtual machine. But it should be fixed now, I tested myself in a Google Colab environment and it worked:

![image](https://github.com/openjournals/joss-reviews/assets/50674829/b20cae6e-4a3a-45d9-a160-99fa8632d867)

I have released a new version of opyrability including this bugfix and it's already available in pip/conda. So your approach of installing in colab should work now (the new version is v1.4.4, and it should be the one automatically downloaded).

Lastly, I have updated the installation instructions at the documentation including your suggestion on how to set up opyrability on a Google Colab virtual environment(https://codes-group.github.io/opyrability/installation.html):

![image](https://github.com/openjournals/joss-reviews/assets/50674829/18513fef-12cf-4b5f-a5d2-7029ea4e0c29)
