### 2023-11-01 #5966 @victoraalves
@kyleniemeyer The minor bugfix that was identified due to @gmxavier's suggestion (see above) made me release a new version for opyrability. I vaguely remember when reading other JOSS reviews that there is a command that can be summoned to the _autobot_ to make it update the version that is written in the first comment (in this case it would bump from 1.4.3 to 1.4.4). How should I proceed here? I am asking because I don't know if this is should be done by me or by you. Thanks in advance!
