+++ README.md
@@ -1,4 +1,4 @@
-# opyrability - Process Operability Analysis in Python.
+# Opyrability - Process Operability Analysis in Python.
 
 Welcome to opyrability, a Python-based package for performing [Process Operability](https://www.sciencedirect.com/science/article/pii/S1474667017338028) analysis.
 
