+++ conda/meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "1.4.3" %}
+{% set version = "1.4.5" %}
 
 package:
   name: opyrability
