+++ environment.yml
@@ -12,5 +12,5 @@ dependencies:
   - jax>=0.4.0
   - cvxopt>=1.2.7
   - cyipopt>=1.2.0
-  - opyrability >= 1.1
+  - opyrability >= 1.4
   - polytope >=0.2.4
