### 2024-02-05 #5966 @victoraalves
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [X]  Double check authors and affiliations (including ORCIDs)
> * [X]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [X]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [X]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [X]  Make sure that the license listed for the archive is the same as the software license.

Hi @kyleniemeyer I have done aforementioned, please see the Zenodo DOI below
[DOI: 10.5281/zenodo.10620959](https://doi.org/10.5281/zenodo.10620959)
