### 2024-02-05 #5966 @kyleniemeyer
@victoraalves just some small edits needed before accepting:
* please change "40's" to "1940s" (no apostrophe) - the 2040s are only 16 years away!
* Change "Prof. Georgakis and coworkers" to just "Georgakis and coworkers"
* Eliminate the acronyms introduced (and only used perhaps once): OI, PSE, CSTR, AOS. Just spell these words out
