+++ paper/paper.md
@@ -57,11 +57,11 @@ yield a process that is designed in a way that its operability capabilities
 are hindered. In other words, because the control objectives were
 not considered early in the design phase, the process itself might be not
 controllable or operable at all. To give some perspective on how challenging this
-problem can be, there are reports dating back to the 40's from well-known authors in the process control field such as Ziegler and Nichols 
+problem can be, there are reports dating back to the 1940s from well-known authors in the process control field such as Ziegler and Nichols 
 [@ziegler1943process] mentioning the importance of interconnecting design and control.
 
 Considering this, the need of quantifying achievability for a general nonlinear
-process naturally arises. The underlying motivation of determining whether it would be possible to measure the operability of a process to simultaneously achieve process design and control objectives led Prof. Georgakis and coworkers [@georgakis00;@vinson00;@siva05;@lima10]
+process naturally arises. The underlying motivation of determining whether it would be possible to measure the operability of a process to simultaneously achieve process design and control objectives led Georgakis and coworkers [@georgakis00;@vinson00;@siva05;@lima10]
 to formally define *process operability*
 and a metric called the *Operability Index (OI)*. The OI was conceptualized as a measure to quantify achievability of nonlinear processes [@vinson00], which was proven to be independent of the control strategy and inventory control layer [@vinson02]. In addition, it allows for the efficient ranking of competing designs and/or control structures [@lima10b] and enables the systematic assessment of operability characteristics under the presence of disturbances. Hence,
 process operability was formalized as *a systematic framework to simultaneously assess design and control objectives early in the conceptual phase of industrial, typically large-scale, and nonlinear chemical processes.*
@@ -83,12 +83,12 @@ operability and achievability of industrial processes earlier in the conceptual 
 eliminates the need for resorting to *ad-hoc*-type solutions to the design and control
 of industrial processes, which are inherently with loss of generality. The use of this framework thus guarantees a solution to the operability problem that
 is generalizable to any process, as long as a mathematical model of the given application is available.
-Hence, the introduction of ``opyrability`` in Python, a widely used and freely available programming language, is a significant advancement in the process operability field. Being open-source and hosted in a community-driven environment, it offers a valuable resource to the process systems engineering (PSE), computational catalysis and material sciences communities that would benefit from operability direct/inverse mappings. This package empowers researchers and practitioners to easily investigate the operability aspects of both emerging and existing large-scale industrial processes. Additionally, on a lab scale, it can aid in the examination of material properties that guide design decisions, such as reactions rate and membrane parameters that would be needed to reach certain product specifications.
+Hence, the introduction of ``opyrability`` in Python, a widely used and freely available programming language, is a significant advancement in the process operability field. Being open-source and hosted in a community-driven environment, it offers a valuable resource to the process systems engineering, computational catalysis and material sciences communities that would benefit from operability direct/inverse mappings. This package empowers researchers and practitioners to easily investigate the operability aspects of both emerging and existing large-scale industrial processes. Additionally, on a lab scale, it can aid in the examination of material properties that guide design decisions, such as reactions rate and membrane parameters that would be needed to reach certain product specifications.
 
[…]
