### 2024-02-05 #5966 @victoraalves
> * AOS

@kyleniemeyer done! The only one I kept was the OI since it's mentioned more than once. All other acronyms have been removed. This is in the main branch already.
