### 2024-02-05 #5966 @victoraalves
> @victoraalves just some small edits needed before accepting:
>
> * please change "40's" to "1940s" (no apostrophe) - the 2040s are only 16 years away!
> * Change "Prof. Georgakis and coworkers" to just "Georgakis and coworkers"
> * Eliminate the acronyms introduced (and only used perhaps once): OI, PSE, CSTR, AOS. Just spell these words out

@kyleniemeyer done! The only one I kept was the OI since it's mentioned more than once. All other acronyms have been removed. This is in the main branch already.
