### 2024-02-06 #5966 @victoraalves
@kyleniemeyer  during proof-reading I found a duplicate link of numpy.org. One of them should by scipy.org. I just fixed this and uploaded to the main branch.
