### 2024-02-06 #5966 @kyleniemeyer
@victoraalves ah, we may need to re-process the acceptance to correct that.

Actually, upon closer look, I missed that you were linking to numpy and scipy. Instead, you should cite the relevant publications (numpy: https://doi.org/10.1038/s41586-020-2649-2,  scipy: 10.1038/s41592-019-0686-2). Can you add citations for these, similar to how you cited IPOPT?
