### 2024-02-06 2f3be29
added numpy, scipy, matplotlib citations.
