+++ paper/references.bib
@@ -517,3 +517,189 @@ author = {Victor Alves and Vitor Gazzaneo and Fernando V. Lima},
 keywords = {Process operability, Machine learning, Kriging, Gaussian process regression, Surrogate modeling},
 abstract = {The objective in this work is to develop a machine learning-based framework for process operability using surrogate responses based on Kriging (also known as Gaussian Process Regression). Currently, the available operability approaches for nonlinear systems are limited by the problem dimensionality that they can address, not being computationally tractable for high-dimensional systems. The proposed approach will use Kriging-based models to substitute the developed first-principles or process simulation-based models. The built surrogate models can generate responses that are comparable to the first-principles nonlinear models in terms of accuracy, while reducing the computational effort. To achieve this goal, a framework for the systematic analysis of highly nonlinear, large-dimensional systems at steady state is developed. The proposed approach is benchmarked against current operability methods and provides a new direction in the process operability field employing Kriging models. Two case studies associated with natural/shale gas conversion are addressed to illustrate the effectiveness of the proposed methods, namely a membrane reactor for direct methane conversion to fuels and chemicals and a natural gas combined cycle power plant. It is shown that the computational time for operability calculations is significantly decreased when using the developed approach, with reductions of up to four orders of magnitude, while the relative errors with respect to the output responses is below 0.3\% for the worst-case scenario considering all cases. This work thus contributes to machine learning formulations and algorithms for process operability to enable the improved design, operations and manufacturing of chemical and energy systems.}
 }
+
+
+@Article{Harris2020,
+author={Harris, Charles R.
+and Millman, K. Jarrod
+and van der Walt, St{\'e}fan J.
+and Gommers, Ralf
+and Virtanen, Pauli
+and Cournapeau, David
+and Wieser, Eric
+and Taylor, Julian
+and Berg, Sebastian
+and Smith, Nathaniel J.
+and Kern, Robert
+and Picus, Matti
+and Hoyer, Stephan
+and van Kerkwijk, Marten H.
+and Brett, Matthew
+and Haldane, Allan
+and del R{\'i}o, Jaime Fern{\'a}ndez
+and Wiebe, Mark
+and Peterson, Pearu
+and G{\'e}rard-Marchant, Pierre
+and Sheppard, Kevin
+and Reddy, Tyler
+and Weckesser, Warren
+and Abbasi, Hameer
+and Gohlke, Christoph
+and Oliphant, Travis E.},
+title={Array programming with NumPy},
+journal={Nature},
+year={2020},
+month={Sep},
+day={01},
+volume={585},
+number={7825},
+pages={357-362},
[…]
