### 2024-02-06 #5966 @victoraalves
@kyleniemeyer Absolutely! Thank you for pointing this out. I just did this to those who apply: numpy, scipy and also matplotlib.
