+++ paper/paper.md
@@ -85,7 +85,7 @@ of industrial processes, which are inherently with loss of generality. The use o
 is generalizable to any process, as long as a mathematical model of the given application is available.
 Hence, the introduction of ``opyrability`` in Python, a widely used and freely available programming language, is a significant advancement in the process operability field. Being open-source and hosted in a community-driven environment, it offers a valuable resource to the process systems engineering, computational catalysis and material sciences communities that would benefit from operability direct/inverse mappings. This package empowers researchers and practitioners to easily investigate the operability aspects of both emerging and existing large-scale industrial processes. Additionally, on a lab scale, it can aid in the examination of material properties that guide design decisions, such as reactions rate and membrane parameters that would be needed to reach certain product specifications.
 
-Moreover, ``opyrability`` is built on well-known and developed packages such as (i) numpy [@Harris2020] and (ii) scipy [@Virtanen2020] for linear algebra and scientific computing; (iii) matplotlib [@Hunter2007]for visualizing the operable regions in 2D/3D; (iv) [cvxopt](https://cvxopt.org/) that allows access to 
+Moreover, ``opyrability`` is built on well-known and developed packages such as (i) numpy [@Harris2020] and (ii) scipy [@Virtanen2020] for linear algebra and scientific computing; (iii) matplotlib [@Hunter2007] for visualizing the operable regions in 2D/3D; (iv) [cvxopt](https://cvxopt.org/) that allows access to 
 [glpk](https://www.gnu.org/software/glpk/) for linear programming; (v) [polytope](https://tulip-control.github.io/polytope/) that enables efficient polytopic calculations; and (vi) [cyipopt](https://cyipopt.readthedocs.io/en/latest/?badge=latest) that allows access to IPOPT [@Wachter2006], a state-of-the-art
 nonlinear programming solver, enabling efficient inverse mapping operations within the operability framework. The inverse mapping task is further extended with full support for automatic differentiation, powered by JAX [@jax2018github]. This effort thus facilitates the 
 dissemination of operability concepts and calculations in process systems engineering and other fields. \autoref{fig:fig1}
