### 2024-02-06 #5966 @victoraalves
> OK, looks good to me—congratulations @victoraalves on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.
>
> Many thanks to @gmxavier and @mustafaalsalmi1999 for reviewing this.

Thank you! I'll definitely sign up as a reviewer!
